refactor: rename uses_cases package to use_case
Fix typo in package name and update all import sites across production
code and tests.
