+++ app/src/main/java/cl/figonzal/lastquakechile/quake_feature/ui/QuakeViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import cl.figonzal.lastquakechile.core.data.remote.ApiError
 import cl.figonzal.lastquakechile.core.data.remote.StatusAPI
 import cl.figonzal.lastquakechile.quake_feature.domain.model.Quake
-import cl.figonzal.lastquakechile.quake_feature.domain.uses_cases.GetQuakesUseCase
+import cl.figonzal.lastquakechile.quake_feature.domain.use_case.GetQuakesUseCase
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
